Changes to autosolve + reformatting

+++ src/feedback.h
@@ -2,28 +2,24 @@
 
 #include <vector>
 
-class Feedback
-{
-  public:
-    Feedback(int blacks, int whites)
-        : m_blacks(blacks),
-          m_whites(whites)
-    {
+class Feedback {
+public:
+    Feedback(long blacks, long whites)
+            : m_blacks(blacks),
+              m_whites(whites) {
     }
 
-    int blacks() const
-    {
+    long blacks() const {
         return m_blacks;
     }
 
-    int whites() const
-    {
+    long whites() const {
         return m_whites;
     }
 
-  private:
-    int m_blacks;
-    int m_whites;
+private:
+    long m_blacks;
+    long m_whites;
 };
 
 extern const std::vector<Feedback> ALL_OUTCOMES;
